5767 - Repository: Reset stale data (#5885)
* 5767 - Repository: Reset stale data

* 5767 - On return, reset

* 5767 - rename tsx -> ts

* 5767 - relative import
